Use pouch to store current user after login.
If a current user already exists, by pass the login page.

Also added a loading page while we check for current user.

We will still want to present the login page if the current user
credentials are stale, but I will implement that as part of general
error handling when an api request is rejected with a 403.