Display individual photo albums.
Now we can click on a photo album and view all of the photos in the album.

There's a download link that doesn't work yet. When it works, we'll be able to
store individual files on disk.

Then we will add the ability to download an entire album to disk.

Finally, we'll add the ability to download all of your troop's photos in one
go.