Download individual files or entire photo albums
Right now we start a download for every file as soon as it is requested. This
will not scales to photo albums with hundreds of downloads.

The solution will be to use the async library to build a small task queue that
we can use to limit the number of simultaneous downloads to a more manageable
number... 3... 10... I have no idea what this number actually will be.